+++ src/main/java/com/argos/proyectocuidadointegral/dto/ProveedorDTO.java
@@ -2,7 +2,7 @@ package com.argos.proyectocuidadointegral.dto;
 
 public class ProveedorDTO {
 
-    private Long idProveedor;
+    private Integer idProveedor;
     private String nombre;
     private String ruc;
     private String telefono;
@@ -11,8 +11,8 @@ public class ProveedorDTO {
 
     public ProveedorDTO() {}
 
-    public Long getIdProveedor() { return idProveedor; }
-    public void setIdProveedor(Long idProveedor) { this.idProveedor = idProveedor; }
+    public Integer getIdProveedor() { return idProveedor; }
+    public void setIdProveedor(Integer idProveedor) { this.idProveedor = idProveedor; }
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
     public String getRuc() { return ruc; }
