+++ src/main/java/com/argos/proyectocuidadointegral/model/Proveedor.java
@@ -10,7 +10,7 @@ public class Proveedor {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_proveedor")
-    private Long idProveedor;
+    private Integer idProveedor;
 
     @Column(name = "nombre", nullable = false, length = 150)
     private String nombre;
@@ -32,8 +32,8 @@ public class Proveedor {
 
     public Proveedor() {}
 
-    public Long getIdProveedor() { return idProveedor; }
-    public void setIdProveedor(Long idProveedor) { this.idProveedor = idProveedor; }
+    public Integer getIdProveedor() { return idProveedor; }
+    public void setIdProveedor(Integer idProveedor) { this.idProveedor = idProveedor; }
     public String getNombre() { return nombre; }
     public void setNombre(String nombre) { this.nombre = nombre; }
     public String getRuc() { return ruc; }
